feat(graphql): implement features to support dashboard automated analysis card
